refactor code, change name of folder to plural
