+++ src/routers/admin/groups.ts
@@ -1,5 +1,5 @@
 import express from 'express';
-import * as groupController from '../../controller/admin/groups';
+import * as groupController from '../../controllers/admin/groups';
 const groupsRouter = express.Router();
 
 groupsRouter.get('/', groupController.index);
