+++ src/tests/fixtures/images/teams/england_text.jpg
